dodany ping i poprawki w README

+++ NBP.sh
@@ -3,10 +3,15 @@
 URL=https://www.nbp.pl/kursy/kursyc.html
 LYNX=$(which lynx)
 TMPFILE=$(mktemp tmpXXXXXX)
-$LYNX -dump $URL > $TMPFILE || (echo "Brak polaczenia z NBP" && exit 1)
+
+response=$(ping -c 1 www.nbp.pl)
+#100% packet loss - nie ma neta lub nbp.pl
+if [[ "$response" == *"100% packet loss"* ]]; then
+  exit 1
+fi
+
+$LYNX -dump $URL > $TMPFILE || exit 1
 
 echo "Kursy walut:"
-sed -n '17,23 s/   //p' $TMPFILE | sed -n '1p; 4p; 6p'
-# Z $TMPFILE wybiera linie 17-23, usuwa niepotrzebne spacje | wypisuje linie 146
-
+sed -n '17,29 s/   //p' $TMPFILE | sed -n '1p; 4p; 6p;'
 rm -f $TMPFILE
